Update Project Config Files
This commit updates the typescript configuration to turn on strict
type-checking options, targetting for es6, node module resolution
and excluding certain directories that are outside of the scope
of the typescript project

The package.json has been modified to remove all unnecessary
packages and add typed packages for development. Additionally,
a build script has been added along with specifying a files for
exported types in the project.