Initial Build and Example
This commit runs an inital build through 'npm run build' and
includes an example usage of the package along with a mock
FLLScore Server built using Python. The server can be run using
'python server.py' in the example folder and the example client
can be run using 'node test.js' in the example folder.