Implement Connect and Add Close
This commit implements the connect function and also adds a close
function to the IScoreClient interface (now renamed to IClient). Also,
the constructor has been changed to optionally accept a name parameter
so the user can specify an id for the client connection with the server.

The example js file has been updated to take advantage of the connect
and close functions.

Because a new function as added to the IClient interface, the version
of the package has been incremented to 0.2.0