Add and Implement Send Last Update
This commit adds a Send Last Update command, returning the last update
as a Date object. A simple getter of the last requested update is also
exposed. Both of these new definitions are used in the example script.

Due to the addition in the IClient interface, the package version has
been bumped to 0.3.0