Add and Implement Send Score
This commit adds and implements the send score functionality of the
client. In doing so, it also adds and exposes a number of typed
definitions for the various data structures returned by the server.
These include information about the schedule, information about each
team, and a single object that contains all team info and schedule info.

The example script has been updated to show a call of send score.

The package version has been increased to 0.4.0 due to the new additions
of the public api.