Implement Automatic Watchdog Pings
This commit adds the ability to parse the "Welcome:" message from the
connect method and set a recurring interval to ping the server within
one watchdog interval if another command has not been sent. If another
command is sent, the recurring interval is reset and starts again.

The example script has been updated to show this feature.