Refactor WebClient and WebServer
This commit serves as a major refactor of websocket communication between
the WebClient and WebServer. None of the methods of IClient are needed
on the WebClient because of the following reasons:

1) The WebClient does not control the connection status to the original
server, thus having a connect function doesn't make sense. The WebClient
should connect immediately to the WebServer and then wait for the WebServer
to connect to the original server and relay data
2) The WebClient shouldn't have functions to force data requests always,
which is what the methods of IClient are intended to do. Instead, the
WebServer should control that and relay its cached value on requests.
3) The WebClient shouldn't have a notion of pinging the original server.
The WebServer handles this connection and is in charge of maintaining
its connection to the original server.

Because of these reasons, it doesn't make sense for the WebClient to
implement the IClient interface. The WebClient instead extends an the
EventEmitter interface and allows the consumer to listen on a couple of
events, while also being able to request the cached values of the WebServer.

The WebServer no longer forwards all data between the original server and
its WebClients. Instead, it serves the cached values on requests and also
updates the WebClients through Events. Values are Lazily loaded at the
first request and further additions are required to retain constant pinging
of new data (at a lower frequency then the heartbeat ping). This will be
addressed later.

Necessary Changes to accommodate these refactors have been made to the
top level exported functions and examples have been updated to use the
new api.