Implement Automatic Retrieval
This commit implements automatic retrieval of new data on a last update
difference. The IClient interface has changed so that sendLastUpdate()
resolves a boolean - true if an update is available. The exact date is
still available through the object, so there is no loss of usage.

The WebProxy now takes advantage of this and emits a server broadcast to
update the score info and last update of any clients connected.