Fix Mixed Message Receive Bug
This commit changes the implementation of Client to fix a bug when
receiving messages from the socket. This bug occurred when multiple writes
are issued at the same time, causing the responses to be sent back-to-back.
The .once() listeners for both receives would capture only the first
command causing one promise would resolve with the other getting rejected.
Further, the second command would be discarded as both of the listeners
would end.

Instead, a private map of callback queues are created for each command.
Upon construction, one data listener is started that controls all
receives from the socket. This listeners breaks commands up by the type
of command, then calls one callback from the right queue. This removes
the multiple .once listeners needed for each method.