More Interface Changes and Add Unit Tests
This commit changes interfaces to remove the individual getters of the
opts interfaces on WebProxy and Client and instead add a getter for the
whole opt interface. The WebProxy now has a IWebProxy Interface to conform
to, so a mock can be created for the upcoming integration tests.

This commit also cleans up the constructors of WebProxy and Client, adding
static methods for handling defaults. The close method now resolves even
when the client is disconnect (works as a no-op).