Test Overhaul
This commit includes overhauls to the testing structure, mainly moving
all tests to run as a single test suite by creating classes and providing
a run method. This allows selective additions/subtractions to the full
testing suite.

The mocha.opts file has been added to remove the long options provided
in the npm script declarations with a single call. Similarly, the
coverage script calls on the already existing cover call instead of
calling the same thing.