Update/Fix Lints
This commit updates the linter setup by first updating dependencies in the package.json, then adding lint scripts to the package.json. This required a separate tsconfig.dev.json to be created. The dev config only adds the test files to the "include" section and is referenced by the linter so linting can be done for both src and test files.

This commit also fixes lints in the test and src files.